requirements.txt
streamlit==1.31.0
python-pptx==0.6.22
anthropic==0.8.0
openai==1.6.0
requests==2.31.0
configparser==6.0.0
